Refactor namespaces and update interfaces and classes
Refactored the `SimpleFlag.Core` namespace to `SimpleFlag.Core.DataSource` for better code organization. Renamed and moved `IDatabaseMigration` to `ISimpleFlagDataSourceMigration` in the new namespace. Updated `SimpleFlagDataSourceOptions`, `SimpleFlagOptionsBuilder`, and `SimpleFlagCustomOptionsBuilder` to use the new interface. Renamed `UseCustomDataSource` method to `UseCustom` for conciseness. Moved `ISimpleFlagDataSource` and added `ISimpleFlagDataSourceMigration` and `SimpleFlagDataSourceOptions` to the new namespace. Added `SimpleFlagCustomExtensions` and `SimpleFlagCustomOptionsBuilder` to `SimpleFlag.Custom` namespace.