Updated properties and null-checks in various classes
In this commit, several changes have been made to improve code robustness and flexibility. In `Todo.cs`, the `Title` property now defaults to an empty string if no value is provided. In `CustomMigrationMetaData.cs`, the `SchemaName` property can now be set to `null`. Two new properties, `SchemaName` and `TablePrefix`, have been added to `SimpleFlagDataSourceOptions.cs`, both of which are nullable and default to an empty string. Lastly, in `SimpleFlagDataSource.cs`, a null-check has been added to the `RunMigration` method to prevent potential `NullReferenceException`.