Refactor solution and update project properties
This commit includes a major refactor of the solution, replacing `SimpleFlaq.AspNetCore` with `SimpleFlag.AspNetCore` and updating project references accordingly. The `IsPackable` property was set to `false` for several projects, preventing them from being packed into a NuGet package. A new `Solution Items` project was added, and `PackageId` and `Description` properties were added to the `SimpleFlag.PostgreSQL` and `SimpleFlag` projects.

The `Directory.Build.props` file was introduced to define common properties for all projects, including setting the `TargetFramework` to `net8.0` and the `Version` to `0.0.1-preview.1`. The `ApplicationContext` property in `CustomVersionTableMetaData` was made nullable, and the `SimpleFlagMigrationOptions` property in `PostgresDataSourceMigration` was initialized.

Null checks were added to the `PostgresDataSourceRepository`, and the `TargetFramework`, `ImplicitUsings`, and `Nullable` properties were removed from certain projects as they are now defined in `Directory.Build.props`. The `SimpleFlagAspNetCoreExtensions` class was added to the `SimpleFlag.AspNetCore` project, and the `net8.0` target framework was set for all projects.

Finally, the `Debug|Any CPU` and `Release|Any CPU` configurations for the `SimpleFlag.AspNetCore` project were added to the solution file, and the corresponding configurations for the `SimpleFlaq.AspNetCore` project were removed. The `SimpleFlaq.AspNetCore` project reference was replaced with a reference to `SimpleFlag.AspNetCore` in `DemoApi.Postgresql.csproj`, and the `using SimpleFlaq.AspNetCore;` statement was replaced with `using SimpleFlag.AspNetCore;` in `Program.cs`.