Reformatted Directory.Build.props and reordered using statements in Program.cs
In Directory.Build.props, the `<PropertyGroup>` tag was removed and then added back in without any changes to the properties within the group. This was likely due to a formatting or indentation change in the source code.

In Program.cs, the `using SimpleFlag.PostgreSQL;` statement was moved from above the `using SimpleFlag.AspNetCore;` statement to below it for better code organization. This change does not affect the functionality of the code.