Add showApiExplorer flag to control API explorer visibility
Added a new boolean property `showApiExplorer` to the `SimpleFlagEndpointOptions` class to control the visibility of API endpoints in the API explorer. Modified the `CreateEndpoint` method in the `SimpleFlagEndpointsHelper` class to accept a `showApiExplorer` parameter and conditionally add metadata based on this flag. Updated the `SimpleFlagEndpointDataSource` class to pass the `showApiExplorer` option from `SimpleFlagEndpointOptions` to the `CreateEndpoint` method. Commented out the `options.showApiExplorer = false;` line in `Program.cs` to use the default value of `true`.